Feeling around the language and semantics

+++ golang/src/display/interfaces.go
@@ -11,10 +11,10 @@ type Composable interface {
 
 // Layout and positioning
 type Layoutable interface {
-	Width(width int)
-	GetWidth() int
-	Height(height int)
-	GetHeight() int
+	Width(width float64)
+	GetWidth() float64
+	Height(height float64)
+	GetHeight() float64
 
 	UpdateState(opts *Opts)
 }
